Move NaluReader into utils
Move NaluReader into utils in preparation for use as a generic bitstream
parsing class. Also add the ability to disable emulation prevention byte
skipping.
