+++ src/codec/h264.rs
@@ -4,7 +4,6 @@
 
 pub mod dpb;
 pub mod nalu;
-pub mod nalu_reader;
 pub mod nalu_writer;
 pub mod parser;
 pub mod picture;
