+++ src/codec/h264/nalu_writer.rs
@@ -185,7 +185,7 @@ impl<W: Write> NaluWriter<W> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::codec::h264::nalu_reader::NaluReader;
+    use crate::utils::NaluReader;
 
     #[test]
     fn simple_bits() {
@@ -238,7 +238,7 @@ mod tests {
             writer.write_ue(5u32).unwrap();
         }
 
-        let mut reader = NaluReader::new(&buf);
+        let mut reader = NaluReader::new(&buf, true);
 
         assert_eq!(reader.read_ue::<u32>().unwrap(), 10);
         assert_eq!(reader.read_se::<i32>().unwrap(), -42);
@@ -254,7 +254,7 @@ mod tests {
             writer.write_ue(50u32).unwrap();
         }
 
-        let mut reader = NaluReader::new(&buf);
+        let mut reader = NaluReader::new(&buf, true);
 
         assert_eq!(reader.read_se::<i32>().unwrap(), 30);
         assert_eq!(reader.read_ue::<u32>().unwrap(), 100);
@@ -274,7 +274,7 @@ mod tests {
             }
             assert_eq!(buf, bitstream);
             {
-                let mut reader = NaluReader::new(&buf);
+                let mut reader = NaluReader::new(&buf, true);
                 for byte in input {
                     assert_eq!(*byte, reader.read_bits::<u8>(8).unwrap());
                 }
